+++ frontend/src/entities/lesson/models/index.ts
@@ -1,4 +1,3 @@
 import "./stores.model";
-import "./loading.model";
 
 export * as lessonModel from "./api.model";